Разделила js на модули
