+++ index.html
@@ -6,7 +6,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кексобукинг</title>
-  <script defer src="js/main.js"></script>
+  <script type="module" defer src="js/main.js"></script>
 </head>
 <body>
